Split db and df utils for file import
